requirements.txt
# python 3.13.2
#
streamlit==1.43.0
streamlit-antd-components==0.3.2
streamlit-pandas-profiling==0.1.3
streamlit-lottie==0.0.5
streamlit-option-menu==0.4.0
streamlit-pdf-viewer==0.0.21
langchain==0.3.19
langchain-community==0.3.18
langchain-anthropic==0.3.10
langchain-google-genai==2.1.1
langchain-ollama==0.3.0
langchain-openai==0.3.8
PyPDF2==3.0.1
python-dotenv==1.0.0
tiktoken==0.9.0
faiss-cpu==1.10.0
openai==1.65.2
markdown
duckduckgo-search
beautifulsoup4==4.13.3
legacy-cgi==2.6.2
